+++ scripts/python/pyproject.toml
@@ -28,8 +28,8 @@ fodt-extract-xml-tag = "fodt.splitter:extract_xml_tag"
 fodt-fix-ignored-keywords = "fodt.fix_ignored:fix_ignored"
 fodt-fix-footer-style = "fodt.fix_footer_style:fix_footer_style"
 fodt-fix-letter-k-footer = "fodt.fix_letter_k_footer:fix_letter_k_footer"
-fodt-gen-kw-uri-map = "fodt.keyword_linker:gen_kw_uri_map_cli"
-fodt-link-keywords = "fodt.keyword_linker2:link_keywords"
+fodt-gen-kw-uri-map = "fodt.keyword_uri_map_generator:gen_kw_uri_map_cli"
+fodt-link-keywords = "fodt.keyword_linker:link_keywords"
 fodt-remove-bookmarks-from-master-styles = "fodt.remove_bookmarks:remove_bookmarks_from_master_styles"
 fodt-remove-chapters = "fodt.splitter:remove_chapters"
 fodt-remove-elements = "fodt.splitter:remove_elements"
